% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfcurve-package.R
\docType{package}
\name{mfcurve-package}
\alias{mfcurve-package}
\title{mfcurve: Multi-Factor Curve Analysis for Grouped Data in R}
\description{
Provides functions for multi-factor curve analysis, replicating and extending the functionality of the Stata mfcurve ado. Includes data preprocessing, statistical testing, and visualization of results with confidence intervals.
}
\author{
\strong{Maintainer}: Maximilian Frank \email{maximilian.frank@psy.lmu.de}

Authors:
\itemize{
  \item Daniel Krähmer \email{Daniel.Kraehmer@soziologie.uni-muenchen.de}
  \item Claudia Weileder \email{maximilian.frank@psy.lmu.de}
}

}
\keyword{internal}
